# docs/swagger.yaml
definitions:
  httpext.JsonError:
    properties:
      error:
        type: string
    type: object
  types.Todo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  title: Todos API
paths:
  /todos:
    get:
      description: Get a Todo list.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Todo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpext.JsonError'
      summary: Get a Todo list
      tags:
      - Todos
    post:
      description: Takes a Todo JSON and store in DB. Return saved JSON.
      parameters:
      - description: Todo JSON
        in: body
        name: Todo
        required: true
        schema:
          $ref: '#/definitions/types.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Todo'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpext.JsonError'
      summary: Store a new Todo
      tags:
      - Todos
    put:
      description: Takes a Todo JSON and update in DB by id. Return saved JSON.
      parameters:
      - description: Todo identifier
        in: path
        name: id
        required: true
        type: string
      - description: Todo JSON
        in: body
        name: Todo
        required: true
        schema:
          $ref: '#/definitions/types.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpext.JsonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpext.JsonError'
      summary: Update a Todo
      tags:
      - Todos
  /todos/{id}:
    delete:
      description: Delete a Todo by id.
      parameters:
      - description: Todo identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpext.JsonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpext.JsonError'
      summary: Delete a Todo
      tags:
      - Todos
    get:
      description: Get a Todo by id.
      parameters:
      - description: Todo identifier
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpext.JsonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpext.JsonError'
      summary: Get a Todo
      tags:
      - Todos
swagger: "2.0"